move themeing related stuff into its own folder
